pyproject.toml
[project]
name = "voice-processing"
version = "0.1.0"
description = "A pipeline to automatically transcribe and summarize audio files using local or API-based LLMs."
readme = "README.md"
requires-python = ">=3.11, <3.13"
dependencies = [
    "openai>=1.88.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "typer[all]>=0.16.0",
    "watchdog>=6.0.0",
]

[project.optional-dependencies]
api = ["openai>=1.88.0"]
local = ["openai-whisper>=20240930"]
dev = ["pytest>=8.4.1"]

[project.scripts]
voice-processing-cli = "voice_processing.cli:app"
voice-processing-watcher = "voice_processing.watcher:app"

[tool.pytest.ini_options]
pythonpath = [".", "src"] 

[tool.setuptools.packages.find]
where = ["src"]


